refactoring settings code
