--- src/components/data/reports/reportColumns.js
@@ -1,58 +1,0 @@
-export default {
-  Overdue: [
-    'borrower.name',
-    'borrower.barcode',
-    'borrowerId',
-    'dueDate',
-    'loanDate',
-    'loanPolicy.name',
-    'loanPolicyId',
-    'loanId',
-    'feeFine',
-    'item.title',
-    'item.materialType.name',
-    'item.status.name',
-    'item.barcode',
-    'item.callNumberComponents.prefix',
-    'item.callNumberComponents.callNumber',
-    'item.callNumberComponents.suffix',
-    'item.volume',
-    'item.enumeration',
-    'item.chronology',
-    'item.copyNumber',
-    'item.contributors',
-    'item.location.name',
-    'item.instanceId',
-    'item.holdingsRecordId',
-    'itemId',
-  ],
-
-  ClaimsReturned: [
-    'borrower.name',
-    'borrower.barcode',
-    'borrowerId',
-    'dueDate',
-    'loanDate',
-    'loanPolicy.name',
-    'loanPolicyId',
-    'loanId',
-    'feeFine',
-    'item.title',
-    'item.materialType.name',
-    'item.status.name',
-    'item.barcode',
-    'item.callNumberComponents.prefix',
-    'item.callNumberComponents.callNumber',
-    'item.callNumberComponents.suffix',
-    'item.volume',
-    'item.enumeration',
-    'item.chronology',
-    'item.copyNumber',
-    'item.contributors',
-    'item.location.name',
-    'item.instanceId',
-    'item.holdingsRecordId',
-    'itemId',
-  ],
-
-};
